Move keywords checking to NVD collector
